Il Disaronno mi ha dato un po' di vena poetica con i commit

+++ api/deleteAccount.php
@@ -1,7 +1,7 @@
 <?php
 
 require_once "../libs/DBaccess.php";
-
+require_once  "/libs/helper.php";
 session_start();
 
 $response = array();
@@ -9,11 +9,13 @@ $response = array();
 //Stream standard
 $input = json_decode(file_get_contents("php://input"), true);
 
-$user = $input["user"];
+$user = cleanFromTags($input["user"] ?? "");
+
+$response['ok'] = false;
 
 $dbaccess = new DBaccess();
 
-if($dbaccess-> getConnection()){
+if($dbaccess-> getConnection() && validateCredentials($user)){
 
     $response['ok'] = $dbaccess-> deleteUser($user);
 }
